python/requirements.txt
flask
pydbus
ansible
cached_property
py-jsonapi==0.3.2b0
url2
